BUG: sparse.linalg: fix overflow in `spsolve` due to 32-bit int usage
Fixed the signature of doubleMalloc and particularly doubleCalloc:
possibly malloc needs to allocate >= 2^31 bytes.

Fixed the signature of doubleMalloc and doubleCalloc in the
implementation file: `SUPERLU_MALLOC` now passes the correct value to
`void *superlu_python_module_malloc(size_t size)` which internally calls
`malloc(size)`.

`work = doubleCalloc(n*nrhs)` now executes for right-hand sides B of
size 46679 by 46680 filled with doubles, and some `(size_t)` casts that
fix proper indexing into buffers such as
`work_col = &work[(size_t)j * (size_t)n]`.

+++ scipy/sparse/linalg/_dsolve/SuperLU/SRC/slu_ddefs.h
@@ -241,8 +241,8 @@ extern void    dSetRWork (int, int, double *, double **, double **);
 extern void    dLUWorkFree (int *, double *, GlobalLU_t *);
 extern int     dLUMemXpand (int, int, MemType, int *, GlobalLU_t *);
 
-extern double  *doubleMalloc(int);
-extern double  *doubleCalloc(int);
+extern double  *doubleMalloc(size_t);
+extern double  *doubleCalloc(size_t);
 extern int     dmemory_usage(const int, const int, const int, const int);
 extern int     dQuerySpace (SuperMatrix *, SuperMatrix *, mem_usage_t *);
 extern int     ilu_dQuerySpace (SuperMatrix *, SuperMatrix *, mem_usage_t *);
